+++ cmd/root.go
@@ -8,7 +8,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var Version = "0.11.8"
+var Version = "0.11.9"
 
 var rootCmd = &cobra.Command{
 	Use:   "aix",